Adjust datatype in migration files

+++ db/migrate/20230213132457_create_food.rb
@@ -2,9 +2,9 @@ class CreateFood < ActiveRecord::Migration[7.0]
   def change
     create_table :foods do |t|
       t.string :name
-      t.decimal :measurement_unit
-      t.decimal :price
-      t.decimal :quantity
+      t.float :measurement_unit
+      t.float :price
+      t.float :quantity
       t.references :user, null: false, foreign_key: true
 
       t.timestamps
